+++ config/dev.php
@@ -4,6 +4,8 @@ return array(
     'app' => array(
         // Trap notice, warnings and error as errors during development
         'errorLevel' => 'notice',
+        // Make sure deprecations are caught during development
+        'deprecation' => 'error',
         // Allow errors to stop execution and show a trace
         'debug' => true,
         // Log as much as possible while developing